ai_devsecops_agent/requirements.txt
openai>=1.0.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.0.0
httpx>=0.25.0
azure-storage-blob>=12.0.0
azure-monitor-query